Change Vec2D::iter to return a 2d-iter impl
This impl yields ((u32, u32), &t) directly as opposed to the previous
implementation using .iter().enumerate(). Moreover, the previous
implementation deviated from the expected API in that enumerate() did
not decorate yielded values with a linear index, but a coordinate index
instead. By changing the definition of iter(), we conform to the
expected interface.